+++ Changelog.md
@@ -10,6 +10,7 @@ OpenCore Changelog
 - Added XCPM CPU power management ACPI table for Intel Alder Lake
 - Updated draw order to avoid graphics tearing in OpenCanopy
 - Fixed handling PCI device paths with logical units in ScanPolicy
+- Added `ReconnectGraphicsOnConnect` option for enabling alternative UEFI graphics drivers
 
 #### v0.7.5
 - Revised OpenLinuxBoot documentation